Don't generate config with duplicate emails
The same email could exist in the history under different names. Skip
emails that have already been added before and stick to the first seen
name.

First seen name should be the most recent name considering we walk
backwards in history, so it has a higher chance of being the correct
name.